EasyFFTs

Documentation for EasyFFTs.

EasyFFTs.dominantfrequencies

EasyFFTs.easyfft

EasyFFTs.easymirror

EasyFFTs.magnitude

EasyFFTs.phase

EasyFFTs.dominantfrequencies — Function

dominantfrequencies(ef, n=5, t=0.1, window=length(ef)//50)

Find the n or fewer dominant frequencies in ef, such that the corresponding magnitude is larger than t times the maximum, and at least window indices away from any larger peaks.

EasyFFTs.easyfft — Function

easyfft(s)
easyfft(s, fs)

Keyword arguments

scalebylength::Bool: determines if the response is scaled by its length. Defaults to true.

Compute the Discrete Fourier Transform (DFT) of the input vector s, scaling by length(s) by default. This function uses FFTW.rfft if s has real elements, and FFTW.fft otherwise.

The output is an EasyFFT object, with fields freq and resp containing the frequences and response respectivly.

See also easymirror to get a symestric spectrum.

Examples

julia> using EasyFFTs

julia> s = sin.(1:5);

julia> ef = easyfft(s)
EasyFFT with 3 samples, showing dominant frequencies f = [0.2, 0.4]

julia> ef.resp
3-element Vector{ComplexF64}:
  0.0587205499074596 + 0.0im
   0.441411013590527 - 0.76819000942203im
 0.23045453212899036 - 0.08137937206396029im

julia> ef = easyfft(s, 0.5)
EasyFFT with 3 samples, showing dominant frequencies f = [0.1, 0.2]

EasyFFTs.easymirror — Function

easymirror(v::AbstractVector)
easymirror(ef::EasyFFT)

Given a one-sided spectrum, return a two-sided version by "mirroring" about 0. This convenience function also ajusts the amplitude of v, or the amplitudes of ef.resp appropriately.

Examples

julia> easymirror(0:3)   # Mirroring the amplitudes
7-element Vector{Float64}:
 1.5
 1.0
 0.5
 0.0
 0.5
 1.0
 1.5

julia> easymirror(fill(1, 4))   # Not halving the zero frequency component
7-element Vector{Float64}:
 0.5
 0.5
 0.5
 1.0
 0.5
 0.5
 0.5

julia> ef = EasyFFTs.EasyFFT([0, 0.2, 0.4], [1, 2, 3]);

julia> easymirror(ef)
EasyFFT with 5 samples, showing dominant frequencies f = [-0.4, 0.4]

EasyFFTs.magnitude — Method

magnitude(ef::EasyFFT)

The absolute values of the response vector.

See also: phase

EasyFFTs.phase — Method

phase(ef::EasyFFT)

The phase of the response vector.

See also: magnitude