EasyFFTs

Documentation for EasyFFTs.

EasyFFTs.EasyFFT

EasyFFTs.domfreq

EasyFFTs.easyfft

EasyFFTs.easymirror

EasyFFTs.finddomfreq

EasyFFTs.magnitude

EasyFFTs.phase

EasyFFTs.phased

EasyFFTs.EasyFFT — Type

EasyFFT(frequencies, response)

A type to hold the response and corresponding frequencies of a discrete fourier transform. Has the fields freq and resp, which can be accessed by dot syntax. Mainly intended to be constructed through easyfft.

EasyFFTs.domfreq — Function

domfreq(ef)
domfreq(ef, n=5, t=0.1, window=length(ef)//50)

Find and return a vector containing the dominant frequencies in ef.

Keyword arguments

n: The maximal of dominant peaks to find. Defaults to 5

t: Minimal magnitude as fraction of maximal magnitude. Defaults to 0.1

window: Minimal difference in index between any larger peak. Defaults to length(ef)//50

See also: finddomfreq

EasyFFTs.easyfft — Function

easyfft(s)
easyfft(s, fs)

Compute the Discrete Fourier Transform (DFT) of the input vector s, scaling by 1/length(s) by default. This function uses FFTW.rfft if s has real elements, and FFTW.fft otherwise.

The output is an EasyFFT object, with fields freq and resp containing the frequences and response respectivly.

Keyword arguments

scalebylength::Bool: determines if the response is scaled by its length. Defaults to true.

See also easymirror to get the full symmetric spectrum of real signals.

Examples

julia> using EasyFFTs

julia> fs = 100;  # 100 samples per second

julia> timestamps = range(0, 1, step = 1/fs);

julia> s = sin.(2π * 2 * timestamps); # sine of frequency = 2 Hz

julia> easyfft(s, fs)
EasyFFT with 51 samples.
Dominant component(s):       
   Frequency  │  Magnitude   
╺━━━━━━━━━━━━━┿━━━━━━━━━━━━━╸
    1.9802    │   0.98461    

julia> easyfft(s)  # `fs` defaults to 1
EasyFFT with 51 samples.
Dominant component(s):       
   Frequency  │  Magnitude   
╺━━━━━━━━━━━━━┿━━━━━━━━━━━━━╸
   0.019802   │   0.98461    
              ╵                      

EasyFFTs.easymirror — Function

easymirror(v::AbstractVector)
easymirror(ef::EasyFFT)

Given a one-sided spectrum, return a two-sided version by "mirroring" about 0. This convenience function also ajusts the amplitude of v, or the amplitudes of ef.resp appropriately.

Examples

julia> using EasyFFTs

julia> fs = 100;  # 100 samples per second

julia> timestamps = range(0, 1, step = 1/fs);

julia> s = sin.(2π * 2 * timestamps); # sine of frequency = 2 Hz

julia> easymirror(ef)
EasyFFT with 101 samples.
Dominant component(s):              ╷              
   Frequency  │  Magnitude   
╺━━━━━━━━━━━━━┿━━━━━━━━━━━━━╸
    -1.9802   │   0.4923     
╶─────────────┼─────────────╴
    1.9802    │   0.4923     

EasyFFTs.finddomfreq — Method

finddomfreq(ef)
finddomfreq(ef; n=5, t=0.1, window=length(ef)//50)

Find and return a vector containing the indices of the dominant frequency components in ef.

This function is used internally in the show method for EasyFFT.

Keyword arguments

n: The maximal of dominant peaks to find. Defaults to 5

t: Minimal magnitude as fraction of maximal magnitude. Defaults to 0.1

window: Minimal difference in index between any larger peak. Defaults to length(ef)//50

See also: domfreq

EasyFFTs.magnitude — Method

magnitude(ef::EasyFFT)

The absolute values of the response vector.

See also: phase, phased

EasyFFTs.phase — Method

phase(ef::EasyFFT)

The phase of the response vector.

See also: magnitude, phased

EasyFFTs.phased — Method

phased(ef::EasyFFT)

The phase of the response vector in degrees.

See also: phase, magnitude